+++ lib/validators/totp-token-validator.js
@@ -1,17 +1,19 @@
+/* jshint camelcase:false */
 
 /**
  * Module dependencies
  */
 
+var Assert = require('validator.js').Assert;
+var Validator = require('validator.js').Validator;
 var Violation = require('validator.js').Violation;
 
 /**
  * Custom phone number validator based on `libphonenumber`.
  *
- * Based on RFC4226 ("Implementations MUST extract a 6-digit
- * code at a minimum and possibly 7 and 8-digit code").
+ * Minimum of 6, maximum 8-digit code.
  *
- * @see https://tools.ietf.org/html/rfc4226
+ * @see http://tools.ietf.org/html/rfc6238
  */
 
 module.exports = function() {
@@ -44,38 +46,18 @@ module.exports = function() {
    */
 
   this.validate = function(value) {
-    var type = typeof value;
-
-    if ('number' !== type && 'string' !== type) {
-      throw new Violation(this, value, { value: 'must_be_a_string_or_number' });
+    if ('string' !== typeof value) {
+      throw new Violation(this, value, { value: Validator.errorCode.must_be_a_string });
     }
 
-    if (!/^(?:(?:0|[1-9][0-9]*))$/.test(value)) {
-      throw new Violation(this, value, { value: 'must_be_an_unsigned_integer' });
+    if (!/^[0-9]+$/.test(value)) {
+      throw new Violation(this, value, { value: 'must_be_numeric' });
     }
 
-    // Generic way of calculating length
-    var length = ('' + value).length;
+    var violation = new Assert().Length(this.boundaries).check(value);
 
-    // Reliable way to calculate length for integers
-    if ('number' === type) {
-      length = (value === 0) ? 1 : parseInt(Math.log(value) / Math.log(10), 10) + 1;
-    }
-
-    if (isNaN(length)) {
-      throw new Violation(this, value, { max: this.boundaries.max });
-    }
-
-    if (this.boundaries.min === this.boundaries.max && length !== this.boundaries.min) {
-      throw new Violation(this, value, { min: this.boundaries.min, max: this.boundaries.max, length: length });
-    }
-
-    if (length > this.boundaries.max) {
-      throw new Violation(this, value, { max: this.boundaries.max, length: length });
-    }
-
-    if (length < this.boundaries.min) {
-      throw new Violation(this, value, { min: this.boundaries.min, length: length });
+    if (true !== violation) {
+      throw violation;
     }
 
     return true;
